// utils/types.ts
export interface User {
    userId: string;
    username: string;
    firstName: string;
    lastName: string;
    email: string; 
    address: string;
    city: string;
    country: string;
    state: string;
    zipCode: string;
    company: string;
    password: string;
    birthdate: Date;
    registeredAt: Date;
    phone: string;
}